requirements.txt
requests==2.28.2            # Biblioteca para realizar requisições HTTP (para a API Football)
python-telegram-bot==20.0   # Biblioteca para interação com o Telegram Bot
python-dotenv==1.0.0        # Para carregar variáveis de ambiente do arquivo .env
